+++ dendrify/__init__.py
@@ -3,5 +3,5 @@ from .ephysproperties import (EphysProperties, default_params,
                               update_default_params)
 from .equations import library
 from .neuronmodel import NeuronModel, PointNeuronModel
-
+from .playground import Playground
 from .version import __version__